Splitting narrator file and integrating into ui
