+++ lib/mumuki/laboratory/locales/narrator.en.yml
@@ -3,9 +3,9 @@ en:
     retry_0: Let's try again!
     retry_1: Keep on!
     retry_2: Let's fix it!
-    introduction_0: "Oops, it didn't work :frowned:."
-    introduction_1: "Oops, it didn't work :frowned:."
-    introduction_2: "Oops, it didn't work :frowned:."
+    introduction_0: "Oops, it didn't work :frowning:."
+    introduction_1: "Oops, it didn't work :frowning:."
+    introduction_2: "Oops, it didn't work :frowning:."
     middle_0: Also, %{tip}.
     middle_1: Also, %{tip}.
     middle_2: Also, %{tip}.
