+++ lib/mumukit/assistant.rb
@@ -1,9 +1,13 @@
 module Mumukit
   class Assistant
+    attr_accessor :rules
+
     def initialize(rules)
-      @rules = rules.map { |it| Mumukit::Assistant::Rule.parse it }
+      @rules = rules
     end
 
+    # Provides tips for the studen for the given submission,
+    # based on the `rules`
     def assist_with(submission)
       @rules
         .select { |it| it.matches?(submission) }
@@ -14,3 +18,4 @@ end
 
 require_relative './assistant/rule'
 require_relative './assistant/message'
+require_relative './assistant/narrator'
